+++ scripts/module.js
@@ -29,6 +29,8 @@ import { setupTemplateHooks } from "./lib/templateHelpers.js";
 import { setupFlourishTracker } from "./lib/flourishTracker.js";
 import { setupAPI } from "./api.js";
 import { setupSkyrimLoadingTips } from "./lib/skyrimLoading.js";
+import { setupMessageUserColor } from "./lib/messageUserColor.js";
+import { setupPanToCombatant } from "./lib/panToCurrentCombatant.js";
 
 export const MODULE_ID = "sundry";
 
@@ -51,6 +53,7 @@ Hooks.once("ready", async function () {
     getSetting("colorize.pf2e-toolbelt.target-helper.roll"),
   );
   setupColorizePersistentPF2eHUD(getSetting("colorize.pf2e-hud.persistent"));
+  setupMessageUserColor(getSetting("colorize.message"));
 
   // Display
   setupDisplayItemPropertyRunes(getSetting("display.item-property-runes"));
@@ -63,6 +66,8 @@ Hooks.once("ready", async function () {
   );
 
   setuplanguageHandling(getSetting("highlight.languages-known"));
+
+  setupPanToCombatant(getSetting("highlight.pan-current-combatant.enabled"));
 
   //Hide
   if (getSetting("hide.sell-all-treasure")) hideSellAllTreasure();