+++ scripts/settings.js
@@ -6,6 +6,7 @@ import { setupDisplayItemPropertyRunes } from "./lib/itemPropertyRunes.js";
 import { setuplanguageHandling } from "./lib/languageHandling.js";
 import { setupNotifySpellstrikeRecharge } from "./lib/notify.js";
 import { openPlayerNotes } from "./lib/openNotes.js";
+import { setupPanToCombatant } from "./lib/panToCurrentCombatant.js";
 import { setupColorizePersistentPF2eHUD } from "./lib/pf2eHUD.js";
 import {
   setupColorizeToolbeltMessageSaves,
@@ -55,6 +56,24 @@ export function setupSettings() {
       },
     },
   );
+
+  
+  game.settings.register(MODULE_ID, "colorize.message", {
+    name: `${MODULE_ID}.module-settings.colorize.message.name`,
+    hint: `${MODULE_ID}.module-settings.colorize.message.hint`,
+    scope: "world",
+    config: true,
+    default: "off",
+    type: String,
+    onChange: (value) => {
+      setupMessageUserColor(value);
+    },
+    choices: {
+      off: `${MODULE_ID}.module-settings.colorize.message.choices.off`,
+      "20-percent": `${MODULE_ID}.module-settings.colorize.message.choices.20-percent`,
+      "set-length": `${MODULE_ID}.module-settings.colorize.message.choices.set-length`,
+    },
+  });
 
   game.settings.register(MODULE_ID, "display.item-property-runes", {
     name: `${MODULE_ID}.module-settings.display.item-property-runes.name`,
@@ -125,6 +144,27 @@ export function setupSettings() {
     },
   });
 
+  game.settings.register(MODULE_ID, "highlight.pan-current-combatant.enabled", {
+    name: `${MODULE_ID}.module-settings.highlight.pan-current-combatant.enabled.name`,
+    hint: `${MODULE_ID}.module-settings.highlight.pan-current-combatant.enabled.hint`,
+    scope: "world",
+    config: true,
+    default: false,
+    type: Boolean,
+    onChange: (value) => {
+      setupPanToCombatant(value);
+    },
+  });
+
+  game.settings.register(MODULE_ID, "highlight.pan-current-combatant.enabled-gm", {
+    name: `${MODULE_ID}.module-settings.highlight.pan-current-combatant.enabled-gm.name`,
+    hint: `${MODULE_ID}.module-settings.highlight.pan-current-combatant.enabled-gm.hint`,
+    scope: "world",
+    config: true,
+    default: false,
+    type: Boolean
+  });
+
   game.settings.register(MODULE_ID, "highlight.loading-tips.items", {
     name: `${MODULE_ID}.module-settings.highlight.loading-tips.items.name`,
     hint: `${MODULE_ID}.module-settings.highlight.loading-tips.items.hint`,
@@ -160,7 +200,7 @@ export function setupSettings() {
     type: Number,
   });
 
-    game.settings.register(MODULE_ID, "highlight.loading-tips.local-last", {
+  game.settings.register(MODULE_ID, "highlight.loading-tips.local-last", {
     name: `${MODULE_ID}.module-settings.highlight.loading-tips.local-last.name`,
     hint: `${MODULE_ID}.module-settings.highlight.loading-tips.local-last.hint`,
     scope: "user",