+++ CHANGELOG.md
@@ -1,9 +1,19 @@
 ## Unreleased
 
-- Add User Color to Message Header
 - **Updated**
   - `Display Property Runes` - Now also displays on character sheet as well as the item description
   - Added Localization where possible for some names
+
+## 1.1.0
+
+- **New**
+  - `Highlight`
+    - `Pan To Current Combatant` - When enabled while in combat it will pan the view to include their token as well as the active combatant (for players only by default)
+  - `Colorize`
+    - `Message Headers` - Adds some styling options to use the user color in the message headers
+- **Updated**
+  - `Colorize PF2e HUD` - Added more colorization for more items
+  - Fixed bug causing `Flourish Tracker` to fail (migrated user checks)
 
 ## 1.0.0